+++ src/lib/store.ts
@@ -1,4 +1,4 @@
-import { Store } from 'tauri-plugin-store';
+import { Store } from "tauri-plugin-store";
 
 // Currently not being used, implement on PostHog maybe?
 const userId = (store: Store) => {
@@ -31,6 +31,16 @@ const theme = (store: Store) => {
   };
 };
 
+const isFirstRun = (store: Store) => {
+  return {
+    get: async () => await store.get<boolean>("isFirstRun"),
+    set: async (value: boolean) => {
+      await store.set("isFirstRun", value);
+      await store.save();
+    },
+  };
+};
+
 const createStore = (path: string) => {
   const store = new Store(path);
 
@@ -38,6 +48,7 @@ const createStore = (path: string) => {
     userId: userId(store),
     windowColor: windowColor(store),
     theme: theme(store),
+    isFirstRun: isFirstRun(store),
   };
 };