+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/).
 
+## 0.9.8-1
+- Add "Used" vs "Free" indicators for footer widgets
+- Set auto-start enabled on first run
+- Add forward and backward history buttons
+
 ## 0.9.7
 - Performance improvements for UI
 - Improve Navigation Header and Sidebar